+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove redundant code in `saw`.
 - Add Julia v1.12 to GitHub workflows.
 - Add Borda count method.
+- Add CRADIS method.
 
 
 ### 0.8.2